+++ extensions/things/src/components/TodoListItemActions.tsx
@@ -16,6 +16,7 @@ import {
   CommandListName,
   Todo,
   setTodoProperty,
+  deleteProject,
   deleteTodo,
   updateTodo,
   updateProject,
@@ -151,18 +152,29 @@ New title:
     await updateAction({ deadline }, { title });
   }
 
-  async function deleteToDo() {
+  async function deleteToDoOrProject() {
+    const isProject = todo.isProject;
+
+    let title = isProject ? 'Delete Project' : 'Delete To-Do';
+    const message = isProject
+      ? 'Are you sure you want to delete this project?'
+      : 'Are you sure you want to delete this to-do?';
+    const deleteFunction = isProject ? deleteProject : deleteTodo;
+
     if (
       await confirmAlert({
-        title: 'Delete To-Do',
-        message: 'Are you sure you want to delete this to-do?',
+        title,
+        message,
         icon: { source: Icon.Trash, tintColor: Color.Red },
       })
     ) {
-      await deleteTodo(todo.id);
+      await deleteFunction(todo.id);
+
+      title = isProject ? 'Deleted project' : 'Deleted to-do';
+
       await showToast({
         style: Toast.Style.Success,
-        title: 'Deleted to-do',
+        title,
         message: todo.name,
       });
       refreshTodos();
@@ -248,7 +260,7 @@ New title:
 
         {lists && lists.length > 0 ? (
           <ActionPanel.Submenu
-            title="Move To"
+            title="Move to"
             icon={Icon.ArrowRight}
             shortcut={{ modifiers: ['cmd', 'shift'], key: 'm' }}
           >
@@ -305,18 +317,18 @@ New title:
           icon={Icon.Trash}
           style={Action.Style.Destructive}
           shortcut={Keyboard.Shortcut.Common.Remove}
-          onAction={deleteToDo}
+          onAction={deleteToDoOrProject}
         />
       </ActionPanel.Section>
 
       {notesURL && (
         <ActionPanel.Section>
           <Action.OpenInBrowser
-            title="Open URL From Notes"
+            title="Open URL from Notes"
             url={notesURL}
             shortcut={{ modifiers: ['cmd', 'shift'], key: 'o' }}
           />
-          <Action.CopyToClipboard title="Copy URL From Notes" content={notesURL} />
+          <Action.CopyToClipboard title="Copy URL from Notes" content={notesURL} />
         </ActionPanel.Section>
       )}