+++ extensions/things/src/tools/add-project.ts
@@ -1,5 +1,5 @@
 import { addProject, ProjectParams } from '../api';
 
 export default async function (props: ProjectParams) {
-  await addProject(props);
+  return await addProject(props);
 }