+++ extensions/things/src/tools/delete-project.ts
@@ -1,0 +1,19 @@
+import { deleteProject, getProjectName } from '../api';
+
+type Input = {
+  /** The project id to delete */
+  projectId: string;
+};
+
+export default async function ({ projectId }: Input) {
+  return await deleteProject(projectId);
+}
+
+export const confirmation = async ({ projectId }: Input) => {
+  const name = await getProjectName(projectId);
+
+  return {
+    message: `Are you sure you want to delete the project?`,
+    info: [{ name: 'Project', value: name }],
+  };
+};